refactor(meet-ai): extract prompt building logic into MeetAiPromptService
