+++ src/meet-ai/services/index.ts
@@ -1,1 +1,2 @@
 export * from './participant-summary.service';
+export * from './meet-ai-prompt.service';
